+++ src/main/java/com/en_circle/slt/plugin/SltWindowFactory.java
@@ -2,6 +2,7 @@ package com.en_circle.slt.plugin;
 
 import com.en_circle.slt.plugin.ui.SltCoreWindow;
 import com.en_circle.slt.plugin.ui.debug.SltDebuggers;
+import com.en_circle.slt.plugin.ui.instance.InstanceInfoPanel;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
@@ -18,6 +19,14 @@ public class SltWindowFactory implements ToolWindowFactory {
             ContentFactory contentFactory = ContentFactory.getInstance();
             Content content = contentFactory.createContent(sltCoreWindow.getContent(),
                     SltBundle.message("slt.ui.process.title"), false);
+            toolWindow.getContentManager().addContent(content);
+        }
+
+        {
+            InstanceInfoPanel infoPanel = new InstanceInfoPanel(toolWindow);
+            ContentFactory contentFactory = ContentFactory.getInstance();
+            Content content = contentFactory.createContent(infoPanel.getContent(),
+                    SltBundle.message("slt.ui.instanceinfo.title"), false);
             toolWindow.getContentManager().addContent(content);
         }