+++ src/main/java/com/en_circle/slt/plugin/swank/SlimeListener.java
@@ -154,6 +154,18 @@ public class SlimeListener implements SwankClient.SwankReply {
             if (request instanceof CompleteSearch completeSearch) {
                 completeSearch.processReply((LispContainer) reply.getItems().get(1));
             }
+
+            if (request instanceof ListThreads listThreads) {
+                listThreads.processReply((LispContainer) reply.getItems().get(1));
+            }
+
+            if (request instanceof KillThread killThread) {
+                killThread.processReply((LispContainer) reply.getItems().get(1));
+            }
+
+            if (request instanceof SuspendThread suspendThread) {
+                suspendThread.processReply((LispContainer) reply.getItems().get(1));
+            }
         } finally {
             requests.remove(replyId.getValue());
         }