+++ src/main/java/com/en_circle/slt/plugin/swank/SwankPacket.java
@@ -264,6 +264,21 @@ public class SwankPacket {
         return new SwankPacket(formatted);
     }
 
+    public static SwankPacket dumpThreads(BigInteger continuation) {
+        String formatted = String.format("(:emacs-rex (swank:list-threads) \":CL-USER\" T %s)", continuation);
+        return new SwankPacket(formatted);
+    }
+
+    public static SwankPacket breakThread(BigInteger thread, BigInteger continuation) {
+        String formatted = String.format("(:emacs-rex (swank:debug-nth-thread %s) \":CL-USER\" T %s)", thread, continuation);
+        return new SwankPacket(formatted);
+    }
+
+    public static SwankPacket killThread(BigInteger thread, BigInteger continuation) {
+        String formatted = String.format("(:emacs-rex (swank:kill-nth-thread %s) \":CL-USER\" T %s)", thread, continuation);
+        return new SwankPacket(formatted);
+    }
+
     private int length;
     private String expressionSource;